# docs/swagger.yaml
definitions:
...
  controller.promotionInput:
    properties:
      createdAt:
        type: string
      isActive:
        type: integer
      productsId:
        type: integer
      promo:
        type: integer
      promoDesc:
        type: string
      promoName:
        type: string
      updatedAt:
        type: string
    type: object
  controller.transactionInput:
    properties:
      createdAt:
        type: string
      name:
        type: string
      payment:
        type: string
      productsId:
        type: integer
      quantity:
        type: integer
      totalPrice:
        type: string
      userId:
        type: integer
    type: object
...
  models.Promotion:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      isActive:
        type: integer
      productsId:
        type: integer
      promo:
        type: integer
      promoDesc:
        type: string
      promoName:
        type: string
      updatedAt:
        type: string
    type: object
  models.Transaction:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      payment:
        type: string
      productsId:
        type: integer
      quantity:
        type: integer
      totalPrice:
        type: string
      userId:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /promotion:
    get:
      description: Get a list of Promotion.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Promotion'
            type: array
      summary: Get all Promotion.
      tags:
      - Promotion
    post:
      description: Creating a new Promotion.
      parameters:
      - description: the body to create a new promotion
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.promotionInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Promotion'
      security:
      - BearerToken: []
      summary: Create New Promotion.
      tags:
      - Promotion
  /promotion/{id}:
    delete:
      description: Delete a Promotion by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: User id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Promotion.
      tags:
      - Promotion
    get:
      description: Get a promo by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: User id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Promotion'
      security:
      - BearerToken: []
      summary: Get promo.
      tags:
      - Promotion
    patch:
      description: Update User by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: User id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update promo
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.promotionInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Promotion'
      security:
      - BearerToken: []
      summary: Update Active Promo.
      tags:
      - Promotion
  /transaction:
    post:
      description: Creating a new Transaction.
      parameters:
      - description: the body to create a new comment
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.transactionInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
      security:
      - BearerToken: []
      summary: Create New Transaction.
      tags:
      - Transaction
  /transaction/{id}:
    get:
      description: Get an Transaction by userId.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: User id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
      security:
      - BearerToken: []
      summary: Get transaction.
      tags:
      - Transaction
  /transaction/user/{id}:
    get:
      description: Get a list of transaction.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Transaction'
            type: array
      security:
      - BearerToken: []
      summary: Get all Transaction By UserId.
      tags:
      - Transaction
  /transation/{id}:
    delete:
      description: Delete a User by user id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: User id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Trannsaction.
      tags:
      - Transaction
...
swagger: "2.0"